Migrate disconnecting logic to v2. Make int tests kill all chrome instance automatically
